feat: JWT authorization, encoded authentication, error code when accessing task

+++ app/api/tasks/route.ts
@@ -1,30 +1,45 @@
 import { database } from "@/database/database"
 import { Task } from "@/models/enum/Task"
 import { NextResponse } from "next/server"
+import { isTokenValid } from "../helpers/jwt"
 
 export async function GET(request: Request): Promise<NextResponse> {
-  const { data, error } = await database.getTasks()
-  if (error) {
-    throw error
+  const { isValid, payload } = isTokenValid(request)
+  if (!isValid) {
+    return NextResponse.json({}, { status: 401 })
   }
+  const { data } = await database.getTasksById(payload.id)
   const response = {
     success: true,
-    data: data
+    data
   }
   return NextResponse.json(response)
 }
 
 export async function POST(request: Request): Promise<NextResponse> {
-  const { user_id } = await request.json()
-  const { data, error } = await database.getTaskByID(user_id)
-  if (error) {
-    throw error
+  const { isValid } = isTokenValid(request)
+  if (!isValid) {
+    return NextResponse.json({}, { status: 401 })
   }
-  const response = { data }
+  
+  const { title, user_id, description, completed } = await request.json()
+  const task: Task = {
+    title,
+    user_id,
+    description,
+    completed,
+  }
+  await database.postTask(task)
+  const response = { success: true }
   return NextResponse.json(response)
 }
 
 export async function PUT(request: Request): Promise<NextResponse> {
+  const { isValid } = isTokenValid(request)
+  if (!isValid) {
+    return NextResponse.json({}, { status: 401 })
+  }
+
   const { id, title, user_id, description, completed } = await request.json()
   const task: Task = {
     id,
@@ -33,24 +48,19 @@ export async function PUT(request: Request): Promise<NextResponse> {
     description,
     completed,
   }
-  const error = await database.putTask(task)
-  if (error) {
-    throw error
-  }
-  const response = {
-    success: true,
-  }
+  await database.putTask(task)
+  const response = { success: true }
   return NextResponse.json(response)
 }
 
 export async function DELETE(request: Request): Promise<NextResponse> {
+  const { isValid } = isTokenValid(request)
+  if (!isValid) {
+    return NextResponse.json({}, { status: 401 })
+  }
+  
   const { id } = await request.json()
-  const error = await database.deleteTask(id)
-  if (error) {
-    throw error
-  }
-  const response = {
-    success: true,
-  }
+  await database.deleteTask(id)
+  const response = { success: true }
   return NextResponse.json(response)
 }
